requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9.11
PyJWT>=2.8
bcrypt
python-dotenv
drf-spectacular

# dev/test
pytest
pytest-django
factory-boy
